requirements.txt
streamlit
pypdf
pymupdf
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
python-dotenv
pandas
Pillow
